// packages/ui/src/molecules/data-grid/hm/types.ts
type TRowBase = Record<string, any>;
export type RowType = 'text' | 'checkbox' | 'date';

export interface IComponentRender {
   DGSelectComponent: FC<RenderCheckboxProps>;
}

export interface ICellRenderProps<TRow extends TRowBase> {
  rowIndex: number;
  // row: TRow;
  row: Flatten<TRow>;
}

// eslint-disable-next-line @typescript-eslint/no-explicit-any
export interface IColumn<TRow extends TRowBase> {
  /** The name of the column. By default it will be displayed in the header cell
   * @see  {mantine-data-grid} https://github.com/Kuechlin/mantine-data-grid/blob/18b1fcc5b2b40f6af6ad9e82d3ee49ccb2111642/docs/pages/examples/AsyncExample.tsx#L72
   * @see  {react-data-grid} https://github.com/Kuechlin/mantine-data-grid/blob/18b1fcc5b2b40f6af6ad9e82d3ee49ccb2111642/docs/pages/examples/AsyncExample.tsx#L78
   * @see  {mantine-data-table} https://github.com/icflorescu/mantine-datatable/blob/main/package/types/DataTableColumn.ts#L17
   */
  name: string;

  /** Render row type    */
  type?: RowType;

  /** A unique key to distinguish each column
   * @see  {mantine-data-grid} https://github.com/Kuechlin/mantine-data-grid/blob/18b1fcc5b2b40f6af6ad9e82d3ee49ccb2111642/docs/pages/examples/AsyncExample.tsx#L72
   * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/src/types.ts#L11
   * @see  {mantine-data-table} https://github.com/icflorescu/mantine-datatable/blob/main/package/types/DataTableColumn.ts#L10
   */
  key: NestedKeyOf<TRow>;

  /** Column width. If not specified, it will be determined automatically
   * based on grid width and specified widths of other columns
   * @see  {mantine-data-grid} https://github.com/Kuechlin/mantine-data-grid/blob/18b1fcc5b2b40f6af6ad9e82d3ee49ccb2111642/docs/pages/examples/AsyncExample.tsx#L80
   * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/website/demos/AllFeatures.tsx#L48
   * @see  {mantine-data-table} https://github.com/icflorescu/mantine-datatable/blob/main/package/types/DataTableColumn.ts#L37
   */
  width?: number | string;

  /** Cell Render props
   * @param idx // row current index
   * @param record // current record object
   * @see  {mantine-data-grid} https://github.com/Kuechlin/mantine-data-grid/blob/main/docs/pages/examples/AsyncExample.tsx#L105
   * ~~~~~~~~~~~~~~~~~~~~~~~~~ https://github.com/TanStack/table/blob/7b9295332306abcc18f23f96be20fe020862c7dd/packages/table-core/src/types.ts#L202
   * ~~~~~~~~~~~~~~~~~~~~~~~~~ https://github.com/TanStack/table/blob/main/packages/table-core/src/core/cell.ts#L4
   * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/src/types.ts#L22
   * @see  {mantine-data-table} https://github.com/icflorescu/mantine-datatable/blob/main/package/types/DataTableColumn.ts#L22
   */
  render?: (props: ICellRenderProps<TRow>) => JSX.Element;

  /** Enables cell editing. If set and no editor property specified,
   *  then a textinput will be used as the cell editor
   * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/src/types.ts#L28
   *
   */
  // editable?: boolean | ((row: TRow) => boolean) | null;

  /** Enable resizing of a column
   * @see  {mantine-data-grid} https://github.com/TanStack/table/blob/7b9295332306abcc18f23f96be20fe020862c7dd/packages/table-core/src/features/ColumnSizing.ts#L51
   * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/src/types.ts#L33
   *
   */
  resizable?: boolean;

  /** Enable sorting of a column
   * @see  {mantine-data-grid} https://github.com/Kuechlin/mantine-data-grid/blob/cac102521eb3f8040be68a9142214e60471757d9/docs/pages/examples/InitialStateExample.tsx#L25
   * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/src/types.ts#L35
   * @see  {mantine-data-table} https://github.com/icflorescu/mantine-datatable/blob/main/package/types/DataTableColumn.ts#L32
   *
   */
  sortable?: boolean;

  // /** Header renderer for the column
  //  * @see  {mantine-data-grid} https://github.com/TanStack/table/blob/main/packages/table-core/src/types.ts#L218
  //  * ~~~~~~~~~~~~~~~~~~~~~~~~~https://tanstack.com/table/v8/docs/examples/react/column-pinning
  //  * @see  {react-data-grid} https://github.com/adazzle/react-data-grid/blob/921fc3d6b6c8b8778af3017aa1688e650a8994d9/src/types.ts#L53
  //  * @see  {mantine-data-table} https://github.com/icflorescu/mantine-datatable/blob/main/package/types/DataTableColumn.ts#L17
  //  *
  //  */
  // headerRenderer?: React.ComponentType<
  //   HeaderRendererProps<TRow, TSummaryRow>
  // > | null;
}

type SortDirection = 'ASC' | 'DESC';

export interface ISortColumn<T extends TRowBase> {
  readonly columnKey: keyof T;
  readonly direction: SortDirection;
}
// ...
export type RDGParserParams<TRow extends TRowBase> = IHMDataGrid<TRow> & {
  setSortColumns: (param: ISortColumn<TRow>[]) => void;
};